Create a matrix of PVT properties for dry- and wet-gas samples

The pvt_gas() generates a table of gas PVT properties at reservoir temperature and pressures from the atmospheric condition up to the initial reservoir pressure. The estimated properties are compressibility factor, formation volume factor, density, compressibility, viscosity, and pseudo-pressure.

pvt_gas(
  input_unit = "Field",
  output_unit = "Field",
  fluid = "wet_gas",
  pvt_model = "DAK",
  visc_model = "Sutton",
  t = 300,
  p = 5000,
  gas_spgr = 0.69,
  nhc_composition = c(0, 0, 0),
  cgr = 3,
  cond_api = 42,
  warning = "yes"
)

Arguments

input_unit
input unit system for parameters, a character string either 'SI' or 'Field'

output_unit
output unit system for properties, a character string either 'SI' or 'Field'

fluid
fluid type, a character string either 'dry_gas' or 'wet_gas'

pvt_model
PVT model, the character string 'DAK'

visc_model
viscosity model, the character string 'Sutton'

t
temperature, a numeric value either in 'C' or 'F' depending on the 'input_unit'

p
pressure, a numeric value either in 'kPag' or 'Psig' depending on the 'input_unit'

gas_spgr
gas specific gravity (Air = 1.0)

nhc_composition
a vector of mole fractions for nitrogen, hydrogen sulfide, and carbon dioxide, respectively

cgr
condensate to gas ratio, a numeric value in 'm3/m3' or 'STB/MMSCF' depending on the 'input_unit'

cond_api
condensate API

warning
a charater string either 'yes' or 'no'

References

Sutton RP (2007). “Fundamental PVT Calculations for Associated and Gas/Condensate Natural-Gas Systems.” SPE Reservoir Evaluation \& Engineering, 10(03), 270--284. ISSN 1094-6470, doi: 10.2118/97099-PA , https://doi.org/10.2118/97099-PA.

Wichert E, Aziz K (1972). “Calculation of Z's for Sour Gases.” Hydrocarbon Processing, 51(5), 119--122.

Examples

pvt_gas_results_1 <- pvt_gas(input_unit = "Field", output_unit = "Field",
fluid = "dry_gas", pvt_model = "DAK", visc_model = "Sutton",
t = 400, p = 20000, gas_spgr = 0.65, nhc_composition = c(0.05,0.02,0.04),
cgr = 0.0, cond_api = NULL, warning = "yes")

#> Warning: pressure is greater than dry gas PVT correlation upper limit: 10000 psig

#> Warning: H2S composition is greater than gas viscosity correlation upper limit: 0.017 mol fraction

head(pvt_gas_results_1)

#>      T_(F) P_(Psig)  Z-Factor Bg_(rb/scf) Density_(lb/cuft) Cg_(1/Psia)
#> [1,]   400        0 0.9996604  0.29451463        0.03000010  0.06806872
#> [2,]   400       10 0.9994315  0.17521848        0.05042543  0.04051521
#> [3,]   400       20 0.9992042  0.12468905        0.07086001  0.02884443
#> [4,]   400       30 0.9989785  0.09677006        0.09130374  0.02239587
#> [5,]   400       40 0.9987546  0.07905999        0.11175651  0.01830521
#> [6,]   400       50 0.9985324  0.06682489        0.13221823  0.01547908
#>      Viscosity_(cp) m(p)_(Psia^2/cp)
#> [1,]     0.01710161             0.00
#> [2,]     0.01710206         23044.84
#> [3,]     0.01710264         57796.51
#> [4,]     0.01710334        104259.22
#> [5,]     0.01710415        162436.93
#> [6,]     0.01710506        232333.32

pvt_gas_results_2 <- pvt_gas(input_unit = "Field", output_unit = "Field",
fluid = "wet_gas", pvt_model = "DAK", visc_model = "Sutton",
t = 300, p = 20000, gas_spgr = 0.75, nhc_composition = c(0.02,0.05,0.08),
cgr = 10.0, cond_api = 42.4, warning = "yes")

#> Warning: pressure is greater than wet gas PVT correlation upper limit: 17000 psig

#> Warning: H2S composition is greater than gas viscosity correlation upper limit: 0.017 mol fraction

head(pvt_gas_results_2)

#>      T_(F) P_(Psig)  Z-Factor Bg_(rb/scf) Density_(lb/cuft) Cg_(1/Psia)
#> [1,]   300        0 0.9991665  0.26012701        0.04087353  0.06810236
#> [2,]   300       10 0.9986016  0.15470785        0.06872508  0.04054888
#> [3,]   300       20 0.9980384  0.11005621        0.09660800  0.02887811
#> [4,]   300       30 0.9974769  0.08538487        0.12452218  0.02242956
#> [5,]   300       40 0.9969172  0.06973491        0.15246753  0.01833892
#> [6,]   300       50 0.9963594  0.05892306        0.18044394  0.01551281
#>      Viscosity_(cp) m(p)_(Psia^2/cp)
#> [1,]     0.01593860             0.00
#> [2,]     0.01593933         24743.52
#> [3,]     0.01594028         62067.98
#> [4,]     0.01594143        111985.57
#> [5,]     0.01594275        174507.97
#> [6,]     0.01594424        249646.48