Create a matrix of PVT properties for black oil samples

The pvt_oil() generates a table of oil and gas PVT properties at reservoir temperature and pressures from the atmospheric condition up to the initial reservoir pressure. The estimated oil properties are solution gas-oil ratio, formation volume factor, density, compressibility, and viscosity. Estimated PVT properties for the associated gas are compressibility factor, formation volume factor, density, compressibility, viscosity, and pseudo-pressure.

pvt_oil(
  input_unit = "SI",
  output_unit = "SI",
  fluid = "black_oil",
  pvt_model = "Standing",
  visc_model = "Beggs_Robinson",
  t = 85.4,
  p = 35000,
  oil_api = 38,
  gas_spgr = 0.67,
  nhc_composition = c(0, 0, 0),
  rsi = NULL,
  pb = 29500,
  warning = "yes"
)

Arguments

input_unit
input unit system for parameters, a character string either 'SI' or 'Field'

output_unit
output unit system for properties, a character string either 'SI' or 'Field'

fluid
fluid type, the character string 'black_oil'

pvt_model
PVT model, a character string. 'Standing', 'Vasquez_Beggs', 'Farshad_Petrosky', 'Al_Marhoun', and 'Glaso' models are currently available

visc_model
viscosity model, a character string. 'Beggs_Robinson', and 'Al_Marhoun' models are currently available

t
temperature, a numeric value either in 'C' or 'F' depending on the 'input_unit'

p
pressure, a numeric value either in 'kPag' or 'Psig' depending on the 'input_unit'

oil_api
API gravity of oil

gas_spgr
gas specific gravity (Air = 1.0)

nhc_composition
a vector of mole fractions for nitrogen, hydrogen sulfide, and carbon dioxide, respectively

rsi
initial solution gas oil ratio in 'm3/m3' or 'SCF/STB' depending on the 'input_unit'. It is either NULL or a numeric value. If 'rsi' is NULL, then a numeric value must be assigned to 'pb'

pb
bubble point pressure, a numeric value either in 'kPag' or 'Psig' depending on the 'input_unit'. it is either NULL or a numeric value. If 'pb' is NULL, then a numeric value must be assigned to 'rsi'

warning
a charater string either 'yes' or 'no'

References

Standing MB (1947). “A Pressure-Volume-Temperature Correlation For Mixtures Of California Oils And Gases.”

Vasquez M, Beggs HD (1980). “Correlations for Fluid Physical Property Prediction.” Journal of Petroleum Technology, 32(06), 968--970. ISSN 0149-2136, doi: 10.2118/6719-PA , https://doi.org/10.2118/6719-PA.

Glaso O (1980). “Generalized Pressure-Volume-Temperature Correlations.” Journal of Petroleum Technology, 32(05), 785--795. ISSN 0149-2136, doi: 10.2118/8016-PA , https://doi.org/10.2118/8016-PA.

Petrosky Jr. GE, Farshad F (1998). “Pressure-Volume-Temperature Correlations for Gulf of Mexico Crude Oils.” SPE Reservoir Evaluation \& Engineering, 1(05), 416--420. ISSN 1094-6470, doi: 10.2118/51395-PA , https://doi.org/10.2118/51395-PA.

Al-Marhoun MA (1988). “PVT Correlations for Middle East Crude Oils.” Journal of Petroleum Technology, 40(05), 650--666. ISSN 0149-2136, doi: 10.2118/13718-PA , https://doi.org/10.2118/13718-PA.

Beggs HD, Robinson JR (1975). “Estimating the Viscosity of Crude Oil Systems.” Journal of Petroleum Technology, 27(09), 1140--1141. ISSN 0149-2136, doi: 10.2118/5434-PA , https://doi.org/10.2118/5434-PA.

Al-Marhoun MA (2004). “Evaluation of empirically derived PVT properties for Middle East crude oils.” Journal of Petroleum Science and Engineering, 42(2), 209--221. ISSN 0920-4105.

Spivey JP, Valko PP, McCain WD (2007). “Applications of the Coefficient of Isothermal Compressibility to Various Reservoir Situations With New Correlations for Each Situation.” SPE Reservoir Evaluation \& Engineering, 10(01), 43--49. ISSN 1094-6470, doi: 10.2118/96415-PA , https://doi.org/10.2118/96415-PA.

Sutton RP (2007). “Fundamental PVT Calculations for Associated and Gas/Condensate Natural-Gas Systems.” SPE Reservoir Evaluation \& Engineering, 10(03), 270--284. ISSN 1094-6470, doi: 10.2118/97099-PA , https://doi.org/10.2118/97099-PA.

Examples

pvt_oil_results_1 <- pvt_oil(input_unit = "Field", output_unit = "Field", fluid = "black_oil",
pvt_model = "Standing", visc_model = "Beggs_Robinson",
t = 200, p = 3000, oil_api = 35, gas_spgr = 0.8,
nhc_composition = c(0.05,0.02,0.04), rsi = 650, pb = NULL, warning = "no")

head(pvt_oil_results_1)

#>      T_(F) P_(Psig) Rso_(scf/stb) Bo_(rb/stb) Oil_Density_(lb/ft3) Co_(1/Psia)
#> [1,]   200        0      4.219422    1.068194             49.66423 0.026686292
#> [2,]   200       10      5.515100    1.068742             49.65195 0.016571877
#> [3,]   200       20      6.865095    1.069314             49.63913 0.012207090
#> [4,]   200       30      8.262028    1.069906             49.62586 0.009750207
#> [5,]   200       40      9.700461    1.070517             49.61218 0.008163618
#> [6,]   200       50     11.176203    1.071144             49.59813 0.007048752
#>      Oil_Viscosity_(cp)  Z-Factor Bg_(rb/scf) Gas_Density_(lb/ft3) Cg_(1/Psia)
#> [1,]           1.738536 0.9983788  0.22570682           0.04817941  0.06815611
#> [2,]           1.724709 0.9972775  0.13416458           0.08105286  0.04060275
#> [3,]           1.710593 0.9961776  0.09539066           0.11399882  0.02893210
#> [4,]           1.696289 0.9950793  0.07396692           0.14701738  0.02248367
#> [5,]           1.681872 0.9939825  0.06037702           0.18010864  0.01839314
#> [6,]           1.667396 0.9928873  0.05098835           0.21327267  0.01556714
#>      Gas_Viscosity_(cp) m(p)_(Psia^2/cp)
#> [1,]         0.01331245             0.00
#> [2,]         0.01331330         29657.74
#> [3,]         0.01331441         74417.23
#> [4,]         0.01331575        134308.46
#> [5,]         0.01331730        209360.76
#> [6,]         0.01331905        299602.81

pvt_oil_results_2 <- pvt_oil(input_unit = "SI", output_unit = "Field", fluid = "black_oil",
pvt_model = "Vasquez_Beggs", visc_model = "Al_Marhoun",
t = 80, p = 20000, oil_api = 34.4, gas_spgr = 0.65,
nhc_composition = c(0.0,0.0,0.0), rsi = NULL, pb = 11350, warning = "yes")

head(pvt_oil_results_2)

#>      T_(F) P_(Psig) Rso_(scf/stb) Bo_(rb/stb) Oil_Density_(lb/ft3) Co_(1/Psia)
#> [1,]   176        0      1.026203    1.068018             49.81744 0.016845054
#> [2,]   176       10      1.900271    1.068433             49.80529 0.011014069
#> [3,]   176       20      2.844981    1.068883             49.79218 0.008329874
#> [4,]   176       30      3.842702    1.069357             49.77834 0.006759855
#> [5,]   176       40      4.883386    1.069852             49.76391 0.005719499
#> [6,]   176       50      5.960453    1.070364             49.74900 0.004974777
#>      Oil_Viscosity_(cp)  Z-Factor Bg_(rb/scf) Gas_Density_(lb/ft3) Cg_(1/Psia)
#> [1,]           1.630904 0.9985323  0.21752863           0.04061749  0.06814562
#> [2,]           1.623309 0.9975356  0.12931689           0.06832416  0.04059222
#> [3,]           1.615100 0.9965406  0.09195366           0.09608609  0.02892153
#> [4,]           1.606481 0.9955473  0.07130938           0.12390330  0.02247306
#> [5,]           1.597574 0.9945557  0.05821394           0.15177582  0.01838249
#> [6,]           1.588459 0.9935658  0.04916688           0.17970366  0.01555645
#>      Gas_Viscosity_(cp) m(p)_(Psia^2/cp)
#> [1,]         0.01244626             0.00
#> [2,]         0.01244687         31715.08
#> [3,]         0.01244767         79575.49
#> [4,]         0.01244863        143610.68
#> [5,]         0.01244974        223849.56
#> [6,]         0.01245100        320320.47