Create a matrix of PVT properties for reservoir water samples

The pvt_water() generates a table of water PVT properties at reservoir temperature and pressures from the atmospheric condition up to the initial reservoir pressure. The estimated water properties are solution gas-water ratio, formation volume factor, density, compressibility, and viscosity.

pvt_water(
  input_unit = "Field",
  output_unit = "Filed",
  fluid = "water",
  pvt_model = "Spivey",
  visc_model = "Spivey",
  t = 220,
  p = 6000,
  salinity = 10,
  gas_saturated = "yes",
  warning = "yes"
)

Arguments

input_unit
input unit system for parameters, a character string either 'SI' or 'Field'.

output_unit
output unit system for properties, a character string

fluid
fluid type, the character string 'water'

pvt_model
PVT model, a character string. 'Spivey', 'Meehan', and 'McCain' models are currently available

visc_model
viscosity model, a character string. 'Spivey', 'Meehan', and 'McCain' models are currently available

t
temperature, a numeric value either in 'C' or 'F' depending on the 'input_unit'

p
pressure, a numeric value either in 'kPag' or 'Psig' depending on the 'input_unit'

salinity
water salinity in weight percent TDS

gas_saturated
a charater string either 'yes' or 'no'

warning
a charater string either 'yes' or 'no'

References

Meehan DN (1980). “Estimating water viscosity at reservoir conditions.” Petroleum Engineer, 35, 117--118.

Meehan DN (1980). “A correlation for water compressibility.” Petroleum Engineer, 56, 125--126.

McCain Jr. WD (1991). “Reservoir-Fluid Property Correlations-State of the Art (includes associated papers 23583 and 23594 ).” SPE Reservoir Engineering, 6(02), 266--272. ISSN 0885-9248, doi: 10.2118/18571-PA , https://doi.org/10.2118/18571-PA.

Spivey JP, McCain WD, North R (2004). “Estimating density, formation volume factor, compressibility, methane solubility, and viscosity for oilfield brines at temperatures from 0 to 275 C, pressures to 200 MPa, and salinities to 5.7 mole/kg.” Journal of Canadian Petroleum Technology, 43(7), 52--61. ISSN 00219487, doi: 10.2118/04-07-05 , https://doi.org/10.2118/04-07-05.

McCain, Jr. WD, Spivey JP, Lenn CP (2011). Petroleum Reservoir Fluid Property Correlations. PennWell Corporation. ISBN 978-1-59370-187-1.

Examples

pvt_water_results_1 <- pvt_water(input_unit = "Field", output_unit = "Field",
fluid = "water", pvt_model = "McCain", visc_model = "McCain",
t = 300, p = 5000, salinity = 10, gas_saturated = "yes", warning = "no")

head(pvt_water_results_1)

#>      T_(F) P_(Psig) Rsw_(scf/stb) Bw_(rb/stb) Density_(lb/ft3)  Cw_(1/Psia)
#> [1,]   300        0     0.8904753    1.079560         61.98275 0.0012082495
#> [2,]   300       10     0.9404562    1.079550         61.98334 0.0007194677
#> [3,]   300       20     0.9903995    1.079539         61.98394 0.0005124380
#> [4,]   300       30     1.0403052    1.079529         61.98454 0.0003980480
#> [5,]   300       40     1.0901732    1.079518         61.98514 0.0003254862
#> [6,]   300       50     1.1400035    1.079508         61.98575 0.0002753565
#>      Viscosity_(cp)
#> [1,]      0.2671543
#> [2,]      0.2672623
#> [3,]      0.2673704
#> [4,]      0.2674787
#> [5,]      0.2675872
#> [6,]      0.2676958

pvt_water_results_2 <- pvt_water(input_unit = "SI", output_unit = "SI",
fluid = "water", pvt_model = "Spivey", visc_model = "Spivey",
t = 100, p = 15000, salinity = 0.0, gas_saturated = "no", warning = "no")

head(pvt_water_results_2)

#>      T_(C)  P_(kPag) Rsw_(rm3/sm3) Bw_(rm3/sm3) Density_(kg/m3)  Cw_(1/kPaa)
#> [1,]   100   0.00000             0     1.042374        958.3693 4.865692e-07
#> [2,]   100  68.94757             0     1.042339        958.4014 4.864637e-07
#> [3,]   100 137.89515             0     1.042304        958.4336 4.863583e-07
#> [4,]   100 206.84272             0     1.042269        958.4657 4.862529e-07
#> [5,]   100 275.79029             0     1.042234        958.4978 4.861476e-07
#> [6,]   100 344.73786             0     1.042199        958.5299 4.860423e-07
#>      Viscosity_(mPa.s)
#> [1,]         0.2810919
#> [2,]         0.2811120
#> [3,]         0.2811321
#> [4,]         0.2811522
#> [5,]         0.2811724
#> [6,]         0.2811925